Add RefreshToken feature [SignUp - SignIn], Modify map [signup - user]

+++ ssms-back/EntityModels/RefreshToken.cs
@@ -9,7 +9,7 @@ namespace SSMS.EntityModels
         {
             IsDeleted = false;
         }
-        public byte TokenId { get; set; }
+        public int TokenId { get; set; }
         public string Token { get; set; }
         public string DeviceInfo { get; set; }
         public string UserId { get; set; }
